@@ -1,20 +1,21 @@
 """Tests for the `cli` module."""
 
+from __future__ import annotations
+
 import pytest
 
 from shellman import cli
 
 
-def test_main():
+def test_main() -> None:
     """Basic CLI test."""
     assert cli.main([]) == 0
 
 
-def test_show_help(capsys):
-    """
-    Show help.
+def test_show_help(capsys: pytest.CaptureFixture) -> None:
+    """Show help.
 
-    Arguments:
+    Parameters:
         capsys: Pytest fixture to capture output.
     """
     with pytest.raises(SystemExit):
